ユーザー関連のファイルをmembersコントローラーに統一
