--- app/views/master/new.html.erb
@@ -1,64 +1,0 @@
-新規ユーザ登録(master権限)
-
-
-
-<%= form_with(model: @user, url: master_index_path, method: :post) do |f| %>
-  <%# エラーメッセージを出す記述 %>
-  <% if @user.errors.any? %>
-      <div style="color: red">
-        <h2>エラーがあります</h2>
-        <ul>
-          <% @user.errors.each do |error| %>
-            <li><%= error.full_message %></li>
-          <% end %>
-        </ul>
-    </div>
-  <% end %>
-  <div class="field">
-    <%= f.label :name %><br />
-    <%= f.text_field :name, autofocus: true %>
-  </div>
-
-  <div class="field">
-    <%= f.label :email %><br />
-    <%= f.email_field :email, autocomplete: "email" %>
-  </div>
-
-  <div class="field">
-    <%= f.label :password %>
-    <% if @minimum_password_length %>
-    <em>(<%= @minimum_password_length %> characters minimum)</em>
-    <% end %><br />
-    <%= f.password_field :password, autocomplete: "new-password" %>
-  </div>
-
-  <div class="field">
-    <%= f.label :password_confirmation %><br />
-    <%= f.password_field :password_confirmation, autocomplete: "new-password" %>
-  </div>
-
-  <div class="field">
-    <p>権限</p>
-    <%= f.label :master %>
-    <%= f.radio_button :role, :master %>
-    <%= f.label :admin %>
-    <%= f.radio_button :role, :admin %>
-    <%= f.label :general %>
-    <%= f.radio_button :role, :general %>
-  </div>
-
-  <div class="field">
-    <p>所属組織</p>
-    <%= f.select :organization_id, options_from_collection_for_select(Organization.all, :id, :name) %>
-  </div>
-
-  <div class="field">
-    <%= f.label :icon %>
-    <%= f.file_field :icon, direct_upload: true %>
-  </div>
-
-  <div class="actions">
-    <%= f.submit "Sign up" %>
-  </div>
-
-<% end %>
