--- app/views/master/show.html.erb
@@ -1,2 +1,0 @@
-<h2>アイコン画像</h2>
-<%= image_tag @user.icon, style: 'width: 100px; height: auto;' %>
