+++ app/views/members/edit.html.erb
@@ -1,6 +1,6 @@
 ユーザー編集
 
-<%= form_with(model: @user, url: master_path, method: :patch) do |f| %>
+<%= form_with(model: @user, url: edit_member_path(@user), method: :patch) do |f| %>
   <%# エラーメッセージを出す記述 %>
   <% if @user.errors.any? %>
       <div style="color: red">
@@ -37,8 +37,11 @@
 
   <div class="field">
     <p>権限</p>
-    <%= f.label :master %>
-    <%= f.radio_button :role, :master %>
+    <%# マスター権限は新規登録を実行するユーザがマスター権限のアカウントであること %>
+    <% if current_user.master? %>
+      <%= f.label :master %>
+      <%= f.radio_button :role, :master %>
+    <% end %>
     <%= f.label :admin %>
     <%= f.radio_button :role, :admin %>
     <%= f.label :general %>
